Add initial handling for Elliptic Curve Keys
This initial patch adds the ability for the creation of TPM resident
Elliptic Curve keys (wrapping comes later).

The keys can only be created with curves that both openssl and the TPM
understand, so this means only prime256v1 and secp384r1 which are NIST
curves.

The TPM will try to verify the hash size based on the hash algorithm,
so we can't do generic signatures (via pkutl) we have to do specific
hash signatures.  The tests, therefore, create an ECDSA self signed
certificate (which extracts the public key) and then verifies the
cert.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>