create_tpm2_key: Add ability to wrap existing ECC Keys
Existing keys are somewhat problematic.  In the pure form, they just
contain all the elliptic curve parameters and we have no idea what the
actual curve is.  Unfortunately the TPM has to know the curve to be
able to import the key, so in this incarnation only process special
keys that contain the curve name.

Subsequently, we'll have to implement logic to identify the curve via
the group parameters so we can import arbitrary ECC keys (although we
still won't be able to import keys whose groups don't correspond to
TPM implemented curves).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>