Remove generic engine key methods
the engine works by attaching to specific engine loaded keys.  The TPM
has no capability for processing generic openssl keys, so adding the
generic methods for passing non-TPM based keys through the TPM is
wrong (and leads to errors).

The way the engine works after this patch is that it *only* attaches
to the key methods of keys specifically loaded through the engine
interface (and validated to be TPM keys).  All other keys will still
be handled by the openssl default routines.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>