Add ability to wrap arbitrary Elliptic Curve Keys
To do this, the TPM still has to have support for the curve, so we
have to translate a generic key to a curve, which can only be done by
comparing the curve parameters with known values.  To enable this we
store the characteristic values of the known curves the TPM supports
and compare against those if EC_GROUP_get_curve_name() fails.

This also allows us to support the bnp256 curve which is unknown to
openssl.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>