Fix errors with multiple keys: Use per connection session directory
The current design means that every new key gets a new connection to
the TPM, which means a separate resource manager (which is per
connection).  However, the key files for volatile keys are now not
unique because of the resource manager virtualizing the handles, so
fix this by creating a unique per connection TPM_DATA_DIR to store all
the handles in.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>