tpm2-common: add public key retrieval when doing parameter encryption
Parameter encryption requires a salt, usually in the form of the
parent key.  To compute the salt parameters the TSS needs access to
the public part of the key.  To ensure this, the TSS keeps files with
public key information around, but the occasional error has been
reported when the public key file gets removed, so ensure it is
available by doing a TPM2_ReadPublic on the salt key before creating a
session.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>