create_tpm2_key: add comment to RSA modulus
I've been asked several times why the RSA code sets the exponent to
zero if it's value is 0x10001.  The reason is that this is a standard
exponent (represented by zero in the TPM) and some TPMs will reject a
value of 0x10001 in its place because they only process the standard
exponent as represented by zero.  Place a comment in the code to try
to make this clearer.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>