e_tpm2-ecc: add decryption key agreement functionality
Elliptic keys don't have a decrypt function like RSA does, which
allows you to pass a session key via encryption to the public key.
Instead they have a Diffie-Hellman Key agreement protocol which allows
agreement on a shared point that can't be deduced by an attacker which
is then used to derive a session key.

Thus, the way EC encryption works is that the sender produces an
ephemeral EC key using the same curve as the recipient's public key.
The sender then computes a point by multiplying the recipient's public
key by the ephemeral private key and computes a symmetric key from this
point via an agreed KDF.  The message then becomes the tuple of
ephemeral public key and encrypted message.  At this point, the sender
can erase the ephemeral key.

The recipient takes the ephemeral public key in the message,
multiplies it by its private key to derive the same shared point and
uses the agreed KDF on the point to compute the symmetric key to
decrypt the message.

Openssl has a somewhat eccentric view of the primitives used by this
operation, but they mostly map to the TPM ECDH_ZGen primitive, so
that's what's added in this case.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>