ecc keys: fix openssl brain damage over named curves
openssl won't actually output a named curve as a puiblic key unless
the EC_GROUP as the asn1_flag OPENSSL_EC_NAMED_CURVE set.  This has
lead to the problem that TPM generated private keys do set named curve
algorithms, but the public keys generated from them are explicitly
parametrised.  Apparently in this case, openssl fails comparison of
the groups for the public and private keys (EVP_PKEY_cmp_parameters)
because part of the comparison is on the group->meth, which is
different because the private key uses an explicit curve method and
the public key uses the generic one.

The main problem this causes is that openssl pkeyutl -derive does a
key comparison, so it will refuse to compute a derived key unless the
input keys are either both explicitly parametrised or both named.

RFC 5480 also requires that curve names be used whenever possible, so
fix this by setting the named curve flag in the engine when we know
the curve.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>