e_tpm2: Don't keep TPM key resident for the openssl key lifetime
Originally e_tpm2 was created as a test for the in-kernel TPM2
resource manager, which means it was supposed to generate situations
where multiple keys were simultaneously in use.  It did this by
loading the TPM version of the key as soon as the openssl key was
loaded and keeping it resident until the openssl key was destroyed
(which can be hours or days depending on the application).  However,
this means that the engine really doesn't work without a resource
manager.  Fix this by making the time the engine keeps the key loaded
as short as possible, making it possible to use the engine on Linux
Kernel versions before 4.12 (the version that introduced the TPM2
resource manager).

Note that this fix is probabalistic: it is still possible for multiple
engine users to try to load keys simultaneously and get a failure
because the TPM is out of resources, but it makes the window where
this can happen much shorter.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>