e_tpm: bug fix: set the TSS directory
Setting the key store directory was a property of the engine prior to

commit a5dab4a412b800462b62092579ccedfc4b7de9e6
Author: James Bottomley <James.Bottomley@HansenPartnership.com>
Date:   Mon Jan 22 14:05:02 2018 -0800

    e_tpm2: Don't keep TPM key resident for the openssl key lifetime

But this was eliminated by the update, so now the TSS tries to write
keys to the current working directory of the application meaning that
if the application starts in an unwriteable location, all engine
actions fail.

Fix this by correctly setting the TPM_DATA_DIR property when the key
is loaded.

Fixes: a5dab4a412b800462b62092579ccedfc4b7de9e6
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>