create_tpm2_key: protect authorization for TPM created keys
Although we use parameter encryption to protect imported keys, the
authorization for the creation of TPM internal keys is currently sent
over the wire in the clear.  Fix this by using a parameter encryption
session (no additional complication because the authority is the first
parameter of TPM2_Create).

Also update the fail_connect.sh test because now create_tpm2_key will
fail to create the session and won't get to the TPM2_Create command.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>