tpm2-common: allow arbitrary hierarchy parents
Instead of hard coding TPM_RH_OWNER, allow this parameter to be passed
in for SRK generation (obviously if it's not TPM_RH_OWNER, it's not
really the srk).  The consumers are now hard coded to TPM_RH_OWNER,
but this is a preparatory move to allowing arbitrary hierarchies for
the key.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>