Allow non-storage hierarchies for key creation
In addition to permanent NV keys, this change also allows the
hierarchy used for seed generation to be explicitly specified by
string.  This means that the parent is now mandatory, however if
parent is missing, TPM_RH_OWNER will be substituted to ensure that
keys created before this change still work as expected.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>