Clean out TPM_DATA_DIR before removing
There's no way to get rid of the name and public bin files for NV
keys.  This leads to unsightly errors when removing the temporary tss
TPM_DATA_DIR because it's non empty, so do a manual removal of the
files belonging to the known NV keys we've been handling.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>