tpm2-common: fix ECC template for primary key
There was still an rsa setting in the unique id.  The TSS fills this
in, but allows you to specify if it's non-zero.  Apparently this
problem causes a build failure on Debian presumably because the
on-stack data area doesn't contain zeros in this location.  Fix this
by making the ECC unique zero by setting both points to zero size.

Fixes: 0af2c22093582a9c6e947fb1842e28dc671f3cea
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>