tests: fix fail_connect.sh on 32 bit platform
It turns out that asn1_integers have a different format on 32 bit
platforms from 64 for unsigned 32 bit integers with the high bit set
(On 64 bit plaforms they occupy 5 bytes and are explicitly unsigned,
on 32 bit platforms they're represented in 4 bytes as a negative
number).  The fail connect test has a hard coded key blob that was
created on a 64 bit platform, so a 32 bit platform doesn't correctly
parse the parent key because it thinks that the integer is too large
and returns -1.  Fix by substituting a 32 bit version of the key blob
(which works on both platforms).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>