Fix parent or hierarchy authorization
The engine contained a hierarchy authorization engine ctl, but it
wasn't correctly plumbed in.  Correctly plumb this in and use
everywhere the parent authorization is required.  Also add a -b option
to create_tpm2_key which allows passing in the parent authorization.

Bonus is that this removes the last vestage of TPM_RS_PW.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>