create_tpm2_key: fix ordering of password collection and policy file parsing
The current order goes collect auth first before parsing the policy
file which leads to the unfortunate consequence that we'll make you
type in a verified password before we error out and tell you you don't
have the correct PolicyAuthValue command in your policy file.  Reorder
this so we only collect the password after we've verified the policy
correctly supports it.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>