fix error handling for failed policy commands
If the policy command of the key fails for any reason, we exit the
engine without flushing the policy session, which leads to a build up
of policy sessions leading to eventual failure.  Fix this by flushing
the policy session in tpm2_init_session() if there's any policy
failure.

Also add a test for this and fix up check_enhanced_auth.sh so that any
failed test exits the script.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>