policy: add the ability to have time limited keys
Implement TPM2_PolicyCounterTimer which allows comparison of the
current counter and timer values.  There is a problem with the policy
file in that traditionally the policy file is simply hashed to make
the final policy, but CounterTimer is different: the necessary
information has to be hashed and then that hash is hashed into the
policy meaning that we have to special case the
TPM_CC_PolicyCounterTimer in the computation of the hash.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>