LICENCE: Change from GPLv2 to LGPLv2.1
There are two problems being solved here: the first is integration with the
OpenSSL licence, which usually requires an exception and also a request from
gnutls to reuse some of the engine code.  The solution to both of these is to
relicense to LGPLv2.1 and then move pieces which might be re-used to a proper
library instead of a loadable object.  This commit is the precursor licence
change.  The signoffs below represent all the authors of copyrightable code as
required by a licence change.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>