move to TCG defined OIDs for the key types
We previously stole an OID in the TCG space to use to identify our
keys.  With the shift to a new key format, we've also taken the
opportunity to clarify with the TCG which OIDs we can use.  It turns
out the ones we initially chose were non-leaf nodes, so we now finally
have a couple of leaf nodes allocated and thus we take advantage of
the new key format to switch to using them.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>