create_tpm2_key: fix name algorithm selection
The results from strcasecmp are zero if it matches, so you need a not
in front of strcasecmp() for the condition to be "it matches".  The
current strncasecmp() in the name algorithm selection were missing the
not's resulting in the wrong algorithm being selected if you specified
it on the command line.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>