check_enhanced_auth.sh: add loop over name algorithm type
This test checks for two things, firstly that we get the policy right
with different name algorithms and secondly that we actually get the
name algorithm correctly specified.  This all works because with
policyPCR the hash of the expected policy registers has to be done
with the name algorithm hash, so we'll get a mismatch here if either
the name algorithm is wrong (or the policy parsing fails).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>