create_tpm2_key: fix use after free
we use pointers to the public and private areas of the TPM
import/create commands to fill out the OpenSSL TPM key, but these
areas go out of scope and are thus freed before we actually use them.
So fix this by declaring the in/out parameters for these commands in
global instead of local scope.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>