tpm2-common: remove interfaces only used by create-tpm2-key
The sensitive to duplicate interface is really only used for key
wrapping, and this is only done at key creation time, not on key
processing, so it has no place in the common code.

When moving to create_tpm2_key.c we change the calling convention to
make it more clear that we're not doing full conversion, merely inner
wrapping the key.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>