Add importable keys
Importable keys are a new type of key that can be created without
access to the receiving TPM.  The design of importable keys is that
you "wrap" the key (encrypt with a derived key) in such a way that it
can only be unwrapped by the TPM local parent key.  To do this
wrapping you only need to know the public key of the parent, which can
be shipped in an ordinary openssl public key file.

The key format of importable keys requires an additional encrypted
secret parameter, which basically encodes the wrapping key using ECDH
so it can be decrypted by the parent on import.

Importable keys are designed to be used everywhere the current
loadable keys are used, the only difference being that the TPM engine
converts the importable key to loadable on first load and thereafter
uses it identically to a loadable key.

The new command line option of importable keys is

create_tpm2_key --import <pubparent> --wrap <private key> tmp.key

Where <pubparent> is the parent public key and <private key> is the
key you are wrapping (obviously since you have no access to the TPM,
this option can only be used to wrap existing private keys).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>