e_tpm-rsa.c: add additional padding types
Up to now we've only handled PKCS1 padding, but since this is being
deprecated by NIST, we need to handle all the others.  It turns out
that the RSA layer of openssl can only really handle OAEP with
mgf1(sha1) or unpadded, so add both of those to the engine.  All other
sophisticated padding in openssl is handled at the pmeth layer and
thus only requires an unpadded key operation at the RSA layer, which
we've now added, so the tpm2 engine should now work for any padding
type supported by openssl.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>