Make removal of key files from the temporary directory explicit
We've been obscuring a bug in tpm2_rm_tssdir() for a while in that we
create a key file for the parent non volatile key but don't remove it
again.  We fixed it up in tpm2_rm_tssdir() by hard coding the removal
of the key file belonging to 81000001.  However, this won't work if we
can have an arbitrary NV parent, so make the key file removal
explicit.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>