Fix compile warning on older GCC
GCC cannot, for each possible value of one variable such as "tssl",
keep track of which of all other variables such as "pubkey" and
"privkey" are uninitialized or not. Some GCC versions just assumes
the variable may be uninitialized if it cannot know. Some more recent
GCC versions seem to take the opposite approach.

Either way, a simple refactoring of the error handling eliminates
the uncertainity by making both the if and else clause definitely
set the variables "pubkey" and "privkey" to an initialized value,
so it now compiles fine.

Signed-off-by: Jonathan Wernberg <jonathaw@axis.com>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>