configure.ac: make location of software TPM configurable
openSUSE has chosen to put the tpm_server binary in /usr/lib/ibmtss
which means the current test programmes can't find it.  Fix this by
creating a new configuration variable TPMSERVER corresponding to the
absolute path of the software TPM.  Make failing to find the software
TPM a non-fatal but significantly warned about problem because it only
affects the running of the tests, not the building of the engine.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>