tpm2-common.c: make openssl handle explicitly parametrised curves correcly
Over the years openssl seems to have changed its handling of elliptic
curves to automatically assume the curve is named unless told
otherwise.  When the engine produces an explicit parametrisation, it
now has to reset the asn1_flag of the key to remove the named curve
setting otherwise the public key ends up being empty.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>