tpm2-common: support loading public key only
Users are slightly perplexed when we ask for a password to read the
public part of the key, since there's no password protected
information in there.  This is because we implement only the private
key load, which always asks for a password if one is required.  This
can be fixed by implementing the engine load public key method.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>