Add support for the Intel TSS
Using the previously created abstrations, insert wrappers for the
Intel TSS.  There are significant annoyances like the IBM constants
are taken from the TCG TPM guides, which all begin TPM_ and the Intel
constants are all taken from the TCG ESAPI guide, which all begin
TPM2_ additionally all the ESAPI commands insist on allocating the
return structures whereas the IBM TSS, following the command guides,
uses on-stack structures.  However, the only real results of this are
a huge set of definitions in the intel-tss.h and pointless allocation
and freeing of structures.

Note that the Intel TSS doesn't export enough of the supporting
routines to build some required capabilites (even though it actually
has them internally), so code providing those are directly copied into
intel-tss.h from the IBM TSS and indicated in comments in the header
file.  Since the IBM TSS is BSD licensed, the resulting code is still
under the project Licence LGPL.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>