Work around Intel TSS Premature Abstraction
Apparently no-one at the TCG read the memo on pointless abstractions,
so they have an internal and an external representation for the TPM
handles.  The really annoying thing is that the two are represented by
the same type in the Intel TSS, so there's no way for the compiler to
check if you've kept everything straight.  The other significant
problem is that the internal representation is only valid for as long
as the context lives, so everything that has a longer lifetime than
the context must be in the external representation.  Finally, just to
trip you up, a small number of commands use the external
representation in the API but no-one will tell you which ones.

Fix this by introducing APIs for transforming handles from internal to
external and back, which are nops on the IBM TSS and add a host of
transforms to the code.  Proof that this is correct is simply that the
Intel TSS works.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>