create_tpm2_key: Add PCR policy option to key creation
The only current way to add policy to a key is to use policymaker to
create it.  This is actually a bit beyond most people, so this is a
patch to simplify that: since most people only want very simple
policies like PCR locking.  This patch allows PCR policy to be
specified as --pcr-lock [hash:]<pcr list>.  This option may be
repeated multiple times to add new banks.  The option will collect the
*current* values of the selected PCRs and lock the key to them (this
means it can't be used for importable keys which are usually created
away from the platform containing the TPM).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>