Add an openssl-3 transform to legacy keys
The engine keys we produce have replaced key methods.  However
openssl-3 by default looks for provider methods and won't use the key
methods unless the keys are "legacy".  The input pkey definitely won't
be legacy so we have to convert it to a legacy key buy assigning the
RSA/EC key back into the pkey.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>