Fix tests for openssl-3
Openssl 3 deprecated rsautil which we use for testing.  This wouldn't
be a problem except they also broke it as well for good measure: it no
longer returns an error code when something fails with the key.  The
test infrastructure relies on this return code, so we have to update
all of the rsautl tests to use pkeyutl instead.

Also the return error for unknown curve changed from "unknown curve"
to "invalid curve" so now we have to look for both.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>