create_tpm2_key: fix uninitialized data problem
pcr_lock is created as an on stack variable and then updated
additively.  On some platforms, the stack isn't zero, so it contains
bogus PCR settings that are never cleared which show up as additional
PCRs in the range the user hadn't requested.  Fix this by zeroing the
entire on-stack structure before using it.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>