unseal_tpm2_key: fix seg fault if key unsealing fails
In some circumstances, key unsealing leads to a segmentation fault on
failure.  The problem is that unseal_tpm2_key passes a NULL ppkey
value into tpm2_load_engine_file() because it doesn't want an EVP_PKEY
returned.  However, there's an error leg in this function that tries
to set this to NULL even if no ppkey was passed in (and thus
segfaults).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>