tpm2-common, tpm2-asn: add new opt for authorized policies
Add a mechanism to keys for using a set of signed policies (only one
of which has to pass to use the key).  The option [4] of the
TSSPRIVKEY ASN is used to store the signed policies and a new type
TSSAUTHPOLICY is added as the container for signed policies.
TSSAUTHPOLICY has a name field, so any policy may be named for easier
location.  Any policy may be added at a later date providing it is
signed by the private <key>.  This addition doesn't have to be done on
the machine where the TPM is.  Although passing the new TSSAUTHPOLICY
is added, no processing is done, the code merely recognizes such a key
but mostly can't use it.  A new function tpm2_add_signed_policy() is
added for key creation routines to use.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>