tpm2-common.c: add processing of signed policy authorizations
This changes the way policies are stored in the key and processed.
Since signed policies are chains of policies which may or may not
work, an additional pols[] array is added to the app_data of the key,
with pols[0] being what the prior policy would be and pols[1+] being
where additional signed policy may be added or removed.  Since a
successful TPM2_PolicyAuthorize() invocation replaces the current
policy hash with one derived from the key and the nonce, it makes no
sense for TPM2_PolicyAuthorize() to be anywhere other than at the
beginning of the standard policy or at the end of each of the signed
policies.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>