tpm2-common: update to allow importable sealed data
The original scheme had a different OID for loadable and importable
keys.  However, this turned out to be unnecessary since we can use the
presence of the optional secret to determine whether the object is
loadable or importable.  For sealed data, we'll use the same OID
(2.23.133.10.1.5) and the presence or absence of secret.  This means
the common routines need to be updated to expect this.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>