name algorithm fixes
The import options don't work with a non-sha256 name algorithm,
primarily because they make the wrong assumptions about which hash
they're supposed to be using for the encryption and hmac key
derivations (it's always the name algorithm of the parent, never the
one of the actual key).  Also fix a missing argument to -n on
seal_tpm2_data and set the name algorithm from the key file on
unseal_tpm2_data.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>