signed_tpm2_policy: fix double free
The routine tpm2_new_signed_policy() frees the key, which contains the
policy, on success or failure meaning the TSSAUTHPOLICY has been freed
after exit from this function and we trigger an error if we try to
free it again.  There's also a double free of the EVP_PKEY in
tpm2_new_signed_policy().

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>