Pad elliptic curve points
Some tests with EC keys occasionally fail with TPM_RC_KEY.  This has
been traced to the x or y points of the public key being representable
in fewer bytes than the order of the curve.  Apparently the MS TPM
emulator contains a check that the size of the points is always the
order of the curve.  Fix this by always padding EC points up to the
curve order.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>