add tests for elliptic curve parent authority
During the construction of the prior elliptic curve refactor patch a
mistake was made whereby the parent authority wasn't passed in.  This
wasn't picked up by any of the current tests, so add a test
specifically using parent authority for elliptic curve keys to prevent
any regressions in this area.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>