doc: update the OID in the RFC to match reality
Gary Lin noticed that the OID in the RFC is missing a digit compared
to what's already in use.  This appears to be an oversight from when
the negotiations with the TCG were onging and the code got updated but
not the RFC.  Update the OID base to 2.23.133.10.1 to reflect what
everyone is actually using.

Cc: Gary Lin <glin@suse.com>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>