seal_tpm2_data: fix importable key creation with Intel TSS
The intel TSS has an internal handle representation which is
habitually used for everything.  However, when the data is created off
TPM, such as for importable keys, the external handle representation
must be used to get the correct handle value.  Fix the import case so
the default handle is external not internal and add a missing parent
string parser.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>