decode_encode: fix the provider recursion problem
Normally providers are selected in order, so specifying

--provider default --provider tpm2

is enough to ensure that the default provider is used for all routines
the tpm2 provider doesn't actually provide.  However, there is an
exception to this ordering in that openssl will search the cache first
before going in provider order, so when the implementation beneath
this provider also needs to use openssl (as all tss's do) then you get
into a situation where the first use of a tpm2 routine will populate
the cache for tpm2 only and then all subsequent uses will find tpm2
before default.  This is a serious recursion problem in the tss
because it uses elliptic curve derivation to obtain the encrypted salt
for sessions but it can't go back around and use this providers
routines for EC because they'll eventually need to derive the
encrypted salt ...

The fix is to artificially populate the cache with the default
provider before the tpm2 also gets added.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>