configure.ac: always use the reference tpm if available
The current formulation of the if clause would use swtpm first if
found and then the ibmswtpm2.  Since the latter is the reference
implementation and the former has deviated from it to the extent that
some tests have to be skipped to prevent spurious failures, always
prefer testing with the reference implementation.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>