tpm2-common: fix for openssl Boolean problems
The ASN.1 standard strictly requires a Boolean to be 0xff for true or
0 for false.  Apparently openssl simply writes the value it was given
without checking (which is usually a C value true or false), so a lot
of our BOOLEANS are ending up with the illegal value 1.  Redo the
setting of this variable to be either 0xff for true or absent for
false (as the standard recommends).

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>