tpm2-common: fix encrypted secret size
It's been reported that the ASN.1 representation of the encrypted
secret is always two bytes too big and zero padded at the end.  This
is because the written parameter isn't reset to zero when the secret
was Marshalled and is in fact left at 2 from a previous use.  Fix by
initializing to zero before marshalling.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>