fix the pkcs12 wrapping test for openssl 3.2
Openssl 3.2 has two problems with this test.  The first is that it
will no longer use a CA certificate unless it has the CA constraint
set and the second is that it seems the pkcs12 required parameters are
incompatible with the local openssl.cnf.  Fix both by not using the
local openssl.cnf when generating non-TPM openssl files and use the
v3_ca section of the global openssl.cnf to add the CA constraint.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>